refactor: preview 선언 위치 수정
